move parse_internal_hostname and CLIENT into connection
